requirements.txt
typer==0.6.1
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
rich==13.7.0 